Dependecy file updated - Moved the database session to the Dependecy file for better organization
